Move parameter validation tests to a spearate file
